Move DatePicker styles into its own file.
