+++ packages/tux/src/components/fields/DatePicker.styles.tsx
@@ -1,0 +1,9 @@
+export const tuxDatePickerStyles = {
+  backgroundColor: '#FFFFFF',
+  backgroundColorOutside: '#FAFAFA',
+  backgroundColorDisabled: '#EFF1F1',
+  backgroundColorSelected: '#3BB172',
+  textColor: '#DCE0E0',
+  textColorSelected: '#FFFFFF',
+  textColorToday: '#3BB172',
+}
