+++ packages/tux/src/components/fields/DatePicker.tsx
@@ -2,7 +2,8 @@ import React, { Component } from 'react'
 import classNames from 'classnames'
 import moment from 'moment'
 import DayPicker, { DateUtils } from 'react-day-picker'
-import { tuxColors, tuxInputStyles, tuxDatePickerStyles } from '../../styles'
+import { tuxColors, tuxInputStyles } from '../../styles'
+import { tuxDatePickerStyles } from './DatePicker.styles'
 import { fade } from '../../utils/color'
 
 const DATE_FORMAT = 'YYYY-MM-DD'
