+++ packages/tux/src/styles.ts
@@ -26,13 +26,3 @@ export const tuxButtonStyles = {
     borderColor: '#188E18',
   }
 }
-
-export const tuxDatePickerStyles = {
-  backgroundColor: '#FFFFFF',
-  backgroundColorOutside: '#FAFAFA',
-  backgroundColorDisabled: '#EFF1F1',
-  backgroundColorSelected: '#3BB172',
-  textColor: '#DCE0E0',
-  textColorSelected: '#FFFFFF',
-  textColorToday: '#3BB172',
-}
